Remove TestConfigurationStore (#288)
